+++ src/commands/constants.ts
@@ -2,6 +2,7 @@ export const OPEN_RESOURCE = 'vscode-sailpoint-identitynow.open-resource';
 export const REMOVE_RESOURCE = 'vscode-sailpoint-identitynow.remove-resource';
 export const REFRESH = 'vscode-sailpoint-identitynow.refresh';
 export const REMOVE_TENANT = 'vscode-sailpoint-identitynow.remove-tenant';
+export const EXPORT_CONFIG = 'vscode-sailpoint-identitynow.export-config';
 export const AGGREGATE = 'vscode-sailpoint-identitynow.aggregate-source';
 export const AGGREGATE_DISABLE_OPTIMIZATION = 'vscode-sailpoint-identitynow.aggregate-source-disable-optimization';
 export const RESET_SOURCE = 'vscode-sailpoint-identitynow.reset-source';