+++ src/models/ConfigQuickPickItem.ts
@@ -1,0 +1,6 @@
+import { QuickPickItem } from "vscode";
+
+export interface ObjectTypeItem extends QuickPickItem {
+    // Select all object types by default
+    objectType: string;
+}